fix builder patterns and currency types

+++ src/main/java/com/brinkcommerce/api/management/order/delivery/model/response/Bonus.java
@@ -3,6 +3,6 @@ package com.brinkcommerce.api.management.order.delivery.model.response;
 public record Bonus(
         String transactionId,
         String reservationId,
-        Double amount,
+        Long amount,
         String currencyCode
 ) {}
